various quirks during MS_DB_TIME_TIMESTAMPTOSTRINGT calls
git-svn-id: http://svn.miranda-ng.org/main/trunk@14265 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/UserInfoEx/src/ctrl_tzcombo.h
@@ -50,7 +50,7 @@ public:
 	virtual void	OnApply(MCONTACT hContact, LPCSTR pszProto);
 	virtual void	OnChangedByUser(WORD wChangedMsg);
 
-	void			GetTime(LPTSTR szTime, WORD cchTime);
+	void			GetTime(LPTSTR szTime, int cchTime);
 };
 
 #endif /* _UI_CTRL_TZ_COMBO_INCLUDE_ */
